Create README.md
For any business , Customer churn is one of the most crucial problem. 
 It is the rate at which clients opt out of purchasing more of a company’s product or services

We are going to analysis an E-commerce dataset.
The data set belongs to ABC online E-commerce company. Company wants to know the customers who are going to churn.
Here, we employ classification algorithm to predict customers that are most likely to churn,

 Using the ML model, companies can know in advance the customers that are most likely to give up the company’s services 
 and therefore, come up with customer retention strategies.

The final model we chose is XGB classifier with hyperparameter tuning and it gave an f1 score of 97.5%.
This also means that the feature vectors available as per our dataset  are more representative of the actual input space. 
We also analyzed the feature importance for the models with scores >90%.